- Restructured methodology logic to 2 callable files: save_objectives.py and select_test_cases.py.
